Demo notebook as good starting point for keras gui
Keras training is started in its own thread. This way the notebook and widgets remain interactive. The model is stopped at the end of the epoch and restarted from the stopping epoch. It is possible to stop the model before finishing an epoch, however, reported loss function values and accuracy will be false.