feature/probabilistic-programming
1) The pdf gives a very short introduction to probabilistics graphical modelling.

2) LinReg by Stan, Pymc3 & Edward notebook

Linear regression is used to demonstrate the differences and properties of the
three most usefull probabilistic programming packages for python as by today.

- PyStan which is an interface to de facto standard package Stan. The package implements
a compiler a BUGS (Bayesian inference Using Gibbs Sampling) like language.

- Pymc3.
This implements number of MCMC samplers for inferencing. In my view this package is developed the
fastest by a large comunity. They have rewritten the whole package by the third time.
Even implemented variational approaches for speed up. It uses Theano backend for CPU / GPU optimization.

- Edward
This package implements a variational expectation maximization on the top of tensorflow.
The schema of the language follows the generative concept of all Bayesian packages.
It was written by Kevin Murphy (author of best selling book in MIT press 2012) and David Blei, the founder of LDA
(  http://www.cs.columbia.edu/~blei/ )

This package promises 5x - 10x speed up in comparison to PyMC3 for the price of some accuracy.

3) Experiments w pymc3 notebook

This example shows how nicely a proper probabilistic model can find the
change point in very a short time course.